requirement.txt
Django==3.2.6
django-active-link==0.1.8
django-ckeditor==6.1.0
django-crispy==1.8.1
django-parler==2.2
environs==9.3.3
gunicorn==20.1.0
Pillow==8.3.1
psycopg2-binary==2.9.1
whitenoise==5.3.0